+++ GeniusTests/Extensions/Genius/Symbol.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 extension Symbol: @retroactive Equatable {
 
-	static let notAvailable = Self(.minus, color: .primary, label: .notAvailable)
+	static let unavailable = Self(.minus, color: .primary, label: .unavailable)
 	static let failed = Self(.xmark, color: .red, label: .failed)
 	static let passed = Self(.checkmark, color: .green, label: .passed)
 	static let unknown = Self(.questionmark, color: .red, label: .unknown)